requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-emit-index-url requirements.in
#
asgiref==3.3.4
    # via django
cffi==1.14.5
    # via oso
dj-database-url==0.5.0
    # via django-heroku
django-appconf==1.0.4
    # via django-fsm-log
django-cors-headers==3.7.0
    # via -r requirements.in
django-extensions==3.1.3
    # via -r requirements.in
django-fsm-admin==1.2.4
    # via -r requirements.in
django-fsm-log==2.0.1
    # via -r requirements.in
django-fsm==2.7.1
    # via
    #   -r requirements.in
    #   django-fsm-admin
    #   django-fsm-log
django-heroku==0.3.1
    # via -r requirements.in
django-model-utils==4.1.1
    # via -r requirements.in
django-ninja==0.13.2
    # via -r requirements.in
django-oso==0.8.0
    # via -r requirements.in
django==3.2.4
    # via
    #   -r requirements.in
    #   django-appconf
    #   django-cors-headers
    #   django-extensions
    #   django-fsm-admin
    #   django-fsm-log
    #   django-heroku
    #   django-model-utils
    #   django-ninja
    #   django-oso
djantic==0.3.0
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
oso==0.12.4
    # via django-oso
psycopg2==2.9.1
    # via django-heroku
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via
    #   -r requirements.in
    #   django-ninja
pytz==2021.1
    # via django
sqlparse==0.4.1
    # via django
typing-extensions==3.10.0.0
    # via pydantic
werkzeug==2.0.1
    # via -r requirements.in
whitenoise==5.2.0
    # via django-heroku

# The following packages are considered to be unsafe in a requirements file:
# setuptools
